#3-rename: PostHashTag를 다시 post 디렉토리로 옮김
